Added template command for trc. Added Readme that contains an example on how to generate keys, TRC, and certs for an example ISD.
